#59: Remove https configuration while we migrate to a new AWS account.  Also get SMILES string SVG generation working when building on Windows.  Build with OpenJDK11 in Travis CI

+++ util/create-smiles-svgs.sh
@@ -12,6 +12,11 @@ if [ ! -f ${DIR}/env.sh ]; then
 fi
 . ${DIR}/env.sh
 
+if ! command -v obabel &>/dev/null ; then
+  echo "obabel is not on your PATH.  Exiting without generating SMILES svg files"
+  exit 0
+fi
+
 rm -fr ${OUTDIR}
 mkdir ${OUTDIR}
